refactor(frontend): extract components from scan details modal
Extract components from ScanDetailsModal for re-use later
